First framework and minima
Set minima up to give a basic framework. There are navigations to sites, a better footer, and other files that need to be edited to allow for other changes.
After this, begin looking at the minima page and adding files to allow for creating other changes. Top list are adding links in text that can go to pages not seen on the header as well as centering text. Then focus can be on pictures or embedded videos.