Framework created
Links are now possible! Making sure to title posts correctly, there are way to link it via home as well as via other link methods. Next is images, but they need to be added via github in the newly created images folder in assets. This is the beginning of making content for the site.